refactor(quests): #556 extract quest catalogue into shared @0xsofia/quests package
Removes the duplicated quest definitions that the extension
(types/questTypes.ts) and explorer (config/quests.ts) had to hand-sync and
which had already drifted (daily-certification.type was signal in explorer,
discovery in extension; claimable flags lived only in the extension).

- New source-only package @0xsofia/quests holds the canonical QUEST_DEFINITIONS,
  QUEST_XP_REWARDS, QuestType, QuestDefinition, SocialPlatform.
- Both apps now re-export from it; questTypes.ts keeps the extension-only
  runtime types (Quest, UserProgress, QuestSystemResult, AtomOperations) and
  quests.ts keeps the explorer-only on-chain glue (predicate/verifier ids,
  title/id maps, level math).
- Canonical values follow the extension (it actively reads type/claimable in
  its quest UI); the explorer only reads id/title/xpReward so behaviour is
  unchanged there.

explorer tsc -b clean + 154 tests pass; extension typecheck clean.
