+++ apps/explorer/package.json
@@ -57,7 +57,8 @@
     "graphql": "^16.9.0",
     "@0xsofia/graphql": "workspace:*",
     "@0xsofia/design-system": "workspace:*",
-    "@0xsofia/taxonomy": "workspace:*"
+    "@0xsofia/taxonomy": "workspace:*",
+    "@0xsofia/quests": "workspace:*"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.5.0",
