+++ packages/quests/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@0xsofia/quests",
+  "description": "Sofia quest catalogue — canonical quest definitions + XP rewards, shared across the extension (mints quest-badge triples) and the explorer (reads them back).",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "main": "./src/index.ts",
+  "module": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./src/index.ts"
+    }
+  },
+  "files": [
+    "src"
+  ],
+  "scripts": {
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "typescript": "^5.9.3"
+  }
+}
